+++ data/chat/src/main/java/com/napzak/market/chat/mapper/chatmessage/ResponseToDomain.kt
@@ -1,7 +1,8 @@
-package com.napzak.market.chat.mapper
+package com.napzak.market.chat.mapper.chatmessage
 
 import com.napzak.market.chat.dto.ChatMessageMetadata
 import com.napzak.market.chat.dto.ChatRealtimeMessage
+import com.napzak.market.chat.dto.MessageItem
 import com.napzak.market.chat.model.ProductBrief
 import com.napzak.market.chat.model.ReceiveMessage
 
@@ -9,7 +10,7 @@ import com.napzak.market.chat.model.ReceiveMessage
  * 실시간 채팅 메시지를 도메인 모델로 변환합니다.
  * @throws IllegalArgumentException 필수 인자들이 null인 경우
  */
-fun ChatRealtimeMessage.toDomain(storeId: Long): ReceiveMessage<*> {
+internal fun ChatRealtimeMessage.toDomain(storeId: Long): ReceiveMessage<*> {
     return when (type) {
         "JOIN" -> toJoin(storeId)
         "LEAVE" -> toLeave(storeId)
@@ -28,7 +29,7 @@ fun ChatRealtimeMessage.toDomain(storeId: Long): ReceiveMessage<*> {
 private fun ChatRealtimeMessage.toText(storeId: Long): ReceiveMessage.Text =
     ReceiveMessage.Text(
         messageId = requireNotNull(messageId),
-        roomId = requireNotNull(roomId),
+        roomId = roomId,
         senderId = requireNotNull(senderId),
         text = content ?: "",
         timeStamp = requireNotNull(createdAt),
@@ -44,7 +45,7 @@ private fun ChatRealtimeMessage.toImage(
 ): ReceiveMessage.Image =
     ReceiveMessage.Image(
         messageId = requireNotNull(messageId),
-        roomId = requireNotNull(roomId),
+        roomId = roomId,
         senderId = requireNotNull(senderId),
         imageUrl = requireNotNull(metadata.imageUrls.firstOrNull()),
         timeStamp = requireNotNull(createdAt),
@@ -59,7 +60,7 @@ private fun ChatRealtimeMessage.toProduct(
 ): ReceiveMessage.Product =
     ReceiveMessage.Product(
         messageId = requireNotNull(messageId),
-        roomId = requireNotNull(roomId),
+        roomId = roomId,
         senderId = requireNotNull(senderId),
         product = ProductBrief(
             productId = metadata.productId,
@@ -83,7 +84,7 @@ private fun ChatRealtimeMessage.toProduct(
 private fun ChatRealtimeMessage.toNotice(content: String): ReceiveMessage.Notice =
     ReceiveMessage.Notice(
         messageId = requireNotNull(messageId),
-        roomId = requireNotNull(roomId),
+        roomId = roomId,
         senderId = senderId,
         notice = content,
         timeStamp = requireNotNull(createdAt),
@@ -92,21 +93,113 @@ private fun ChatRealtimeMessage.toNotice(content: String): ReceiveMessage.Notice
 private fun ChatRealtimeMessage.toDate(metadata: ChatMessageMetadata.Date): ReceiveMessage.Date =
     ReceiveMessage.Date(
         messageId = requireNotNull(messageId),
-        roomId = requireNotNull(roomId),
+        roomId = roomId,
         date = metadata.date,
         timeStamp = requireNotNull(createdAt),
     )
 
 private fun ChatRealtimeMessage.toJoin(storeId: Long): ReceiveMessage.Join =
     ReceiveMessage.Join(
-        roomId = requireNotNull(roomId),
+        roomId = roomId,
         senderId = requireNotNull(senderId),
         isMessageOwner = storeId == senderId,
     )
 
 private fun ChatRealtimeMessage.toLeave(storeId: Long): ReceiveMessage.Leave =
     ReceiveMessage.Leave(
-        roomId = requireNotNull(roomId),
+        roomId = roomId,
         senderId = requireNotNull(senderId),
         isMessageOwner = storeId == senderId,
+    )
+
+/**
+ * 채팅 메시지 목록 조회 응답을 도메인 모델로 변환합니다.
+ */
+internal fun MessageItem.toDomain(roomId: Long): ReceiveMessage<*> {
+    return when (metadata) {
+        null -> toText(roomId)
+        is ChatMessageMetadata.Image -> toImage(roomId, metadata)
+        is ChatMessageMetadata.Product -> toProduct(roomId, metadata)
+        is ChatMessageMetadata.EXIT -> toNotice(roomId, metadata.content)
+        is ChatMessageMetadata.REPORTED -> toNotice(roomId, metadata.content)
+        is ChatMessageMetadata.WITHDRAWN -> toNotice(roomId, metadata.content)
+        is ChatMessageMetadata.Date -> toDate(roomId, metadata)
+    }
+}
+
+private fun MessageItem.toText(roomId: Long): ReceiveMessage.Text =
+    ReceiveMessage.Text(
+        messageId = requireNotNull(messageId),
+        roomId = roomId,
+        senderId = requireNotNull(senderId),
+        text = content ?: "",
+        timeStamp = requireNotNull(createdAt),
+        isRead = isRead ?: false,
+        isMessageOwner = isMessageOwner ?: false,
+        isProfileNeeded = isProfileNeeded ?: false,
+    )
+
+private fun MessageItem.toImage(
+    roomId: Long,
+    metadata: ChatMessageMetadata.Image
+): ReceiveMessage.Image =
+    ReceiveMessage.Image(
+        messageId = requireNotNull(messageId),
+        roomId = roomId,
+        senderId = requireNotNull(senderId),
+        imageUrl = requireNotNull(metadata.imageUrls.firstOrNull { !it.isNullOrBlank() }),
+        timeStamp = requireNotNull(createdAt),
+        isRead = isRead ?: false,
+        isMessageOwner = isMessageOwner ?: false,
+        isProfileNeeded = isProfileNeeded ?: false,
+    )
+
+private fun MessageItem.toProduct(
+    roomId: Long,
+    metadata: ChatMessageMetadata.Product
+): ReceiveMessage.Product =
+    ReceiveMessage.Product(
+        roomId = roomId,
+        messageId = messageId ?: throw IllegalArgumentException(),
+        senderId = requireNotNull(senderId),
+        product = ProductBrief(
+            productId = metadata.productId,
+            tradeType = metadata.tradeType,
+            title = metadata.title,
+            price = metadata.price,
+            genreName = metadata.genreName,
+            isProductDeleted = metadata.isProductDeleted ?: false,
+            // 사용되지 않는 속성들
+            isPriceNegotiable = false,
+            photo = "",
+            productOwnerId = 0,
+            isMyProduct = false,
+        ),
+        timeStamp = requireNotNull(createdAt),
+        isRead = isRead ?: false,
+        isMessageOwner = isMessageOwner ?: false,
+        isProfileNeeded = isProfileNeeded ?: false,
+    )
+
+private fun MessageItem.toNotice(
+    roomId: Long,
+    content: String,
+): ReceiveMessage.Notice =
+    ReceiveMessage.Notice(
+        messageId = requireNotNull(messageId),
+        roomId = roomId,
+        senderId = senderId,
+        notice = content,
+        timeStamp = requireNotNull(createdAt),
+    )
+
+private fun MessageItem.toDate(
+    roomId: Long,
+    metadata: ChatMessageMetadata.Date
+): ReceiveMessage.Date =
+    ReceiveMessage.Date(
+        messageId = requireNotNull(messageId),
+        roomId = roomId,
+        date = metadata.date,
+        timeStamp = requireNotNull(createdAt),
     )
