+++ data/chat/src/main/java/com/napzak/market/chat/mapper/chatroom/EntityToDomain.kt
@@ -1,9 +1,10 @@
 package com.napzak.market.chat.mapper.chatroom
 
-import com.napzak.market.chat.mapper.toDomain
 import com.napzak.market.chat.model.ChatRoom
 import com.napzak.market.chat.model.ChatRoomInformation
+import com.napzak.market.chat.model.ProductBrief
 import com.napzak.market.chat.model.StoreBrief
+import com.napzak.market.local.room.entity.ChatProductEntity
 import com.napzak.market.local.room.entity.ChatRoomEntity
 import com.napzak.market.local.room.relation.ChatRoomWithProduct
 
@@ -35,4 +36,17 @@ private fun ChatRoomEntity.toStoreBrief() = StoreBrief(
     isReported = isReported,
     isOpponentStoreBlocked = isOpponentStoreBlocked,
     isMyStoreBlocked = isChatBlocked,
+)
+
+internal fun ChatProductEntity.toDomain(): ProductBrief = ProductBrief(
+    productId = productId,
+    tradeType = tradeType,
+    title = title,
+    price = price,
+    genreName = genreName,
+    isProductDeleted = isProductDeleted,
+    isPriceNegotiable = isPriceNegotiable,
+    photo = photo ?: "",
+    productOwnerId = productOwnerId ?: 0L,
+    isMyProduct = isMyProduct,
 )
