+++ domain/chat/src/main/java/com/napzak/market/chat/repository/ChatSocketRepository.kt
@@ -1,7 +1,8 @@
 package com.napzak.market.chat.repository
 
+import com.napzak.market.chat.model.ReceiveMessage
 import com.napzak.market.chat.model.SendMessage
-import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.flow.Flow
 
 interface ChatSocketRepository {
     suspend fun connect(): Result<Unit>
@@ -9,6 +10,6 @@ interface ChatSocketRepository {
     suspend fun subscribeChatRoom(roomId: Long): Result<Unit>
     suspend fun subscribeCreateChatRoom(storeId: Long): Result<Unit>
     suspend fun sendChat(chat: SendMessage<*>): Result<Unit>
-    fun collectMessagesFromSocket(storeId: Long, coroutineScope: CoroutineScope)
-    fun collectNewChatRequestFromSocket(coroutineScope: CoroutineScope)
+    fun getChatMessageStream(storeId: Long): Flow<ReceiveMessage<*>>
+    fun getNewChatRequestStream(): Flow<Long>
 }
