+++ domain/chat/src/main/java/com/napzak/market/chat/usecase/ConnectChatSocketUseCase.kt
@@ -8,12 +8,6 @@ class ConnectChatSocketUseCase @Inject constructor(
     private val chatSocketRepository: ChatSocketRepository,
 ) {
     suspend operator fun invoke(storeId: Long, coroutineScope: CoroutineScope): Result<Unit> {
-        return with(chatSocketRepository) {
-            connect()
-                .onSuccess {
-                    collectMessagesFromSocket(storeId, coroutineScope)
-                    collectNewChatRequestFromSocket(coroutineScope)
-                }
-        }
+        return chatSocketRepository.connect()
     }
 }
