+++ feature/main/src/main/java/com/napzak/market/main/WebSocketLifecycleObserver.kt
@@ -4,6 +4,8 @@ import androidx.lifecycle.DefaultLifecycleObserver
 import androidx.lifecycle.LifecycleOwner
 import com.napzak.market.chat.usecase.ConnectChatSocketUseCase
 import com.napzak.market.chat.usecase.DisconnectChatSocketUseCase
+import com.napzak.market.chat.usecase.HandleChatMessageStreamUseCase
+import com.napzak.market.chat.usecase.HandleNewChatRequestStreamUseCase
 import com.napzak.market.chat.usecase.SubscribeChatRoomsUseCase
 import com.napzak.market.store.model.StoreInfo
 import com.napzak.market.store.repository.StoreRepository
@@ -24,10 +26,14 @@ class WebSocketLifecycleObserver @Inject constructor(
     private val tokenProvider: TokenProvider,
     private val connectChatSocketUseCase: ConnectChatSocketUseCase,
     private val disconnectChatSocketUseCase: DisconnectChatSocketUseCase,
+    private val handleChatMessageStreamUseCase: HandleChatMessageStreamUseCase,
+    private val handleNewChatRequestStreamUseCase: HandleNewChatRequestStreamUseCase,
     private val subscribeChatRoomsUseCase: SubscribeChatRoomsUseCase,
 ) : DefaultLifecycleObserver {
     private lateinit var activityScope: CoroutineScope
     private var loginStateCollectJob: Job? = null
+    private var messageCollectJob: Job? = null
+    private var newChatRequestCollectJob: Job? = null
     private val isLoggedIn = MutableStateFlow(false)
 
     override fun onCreate(owner: LifecycleOwner) {
@@ -44,6 +50,9 @@ class WebSocketLifecycleObserver @Inject constructor(
                     runCatching {
                         connectChatSocket(storeId)
                         subscribeChatRooms(storeId)
+                    }.onSuccess {
+                        collectMessages(storeId)
+                        collectNewChatRequests()
                     }
                 }
             }
@@ -53,6 +62,8 @@ class WebSocketLifecycleObserver @Inject constructor(
     override fun onPause(owner: LifecycleOwner) {
         super.onPause(owner)
         loginStateCollectJob?.cancel()
+        messageCollectJob?.cancel()
+        newChatRequestCollectJob?.cancel()
         activityScope.launch { disconnectChatSocketUseCase() }
     }
 
@@ -75,6 +86,18 @@ class WebSocketLifecycleObserver @Inject constructor(
 
     private suspend fun subscribeChatRooms(storeId: Long) {
         subscribeChatRoomsUseCase(storeId = storeId)
+    }
+
+    private fun collectMessages(storeId: Long) {
+        messageCollectJob = activityScope.launch {
+            handleChatMessageStreamUseCase(storeId)
+        }
+    }
+
+    private fun collectNewChatRequests() {
+        newChatRequestCollectJob = activityScope.launch {
+            handleNewChatRequestStreamUseCase()
+        }
     }
 
     private suspend fun isTokenAvailable() = tokenProvider.getAccessToken() != null
